requirements.txt
aiofiles==0.8.0 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.9.5 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
aiosqlite==0.17.0 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
asyncpraw==7.7.1 ; python_version >= "3.12" and python_version < "4.0"
asyncprawcore==2.4.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.0.0 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.20.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-yaml==1.3.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
ruamel-yaml-clib==0.2.8 ; platform_python_implementation == "CPython" and python_version < "3.13" and python_version >= "3.12"
ruamel-yaml==0.18.6 ; python_version >= "3.12" and python_version < "4.0"
sentry-sdk==2.10.0 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy[asyncio]==2.0.31 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
update-checker==0.18.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
yarl==1.9.4 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.19.2 ; python_version >= "3.12" and python_version < "4.0"
